service1/app/requirements.txt
fastapi==0.68.0
pydantic==1.9.0
dapr-ext-fastapi==1.10.0